UPDATED: shortcutsDialog with latest shortcuts RENAMED: MenuAccelerators -> WithMenuAccelerators UPDATED: Windows/Linux about dialog title
